data-collector: Extracted reusable logic from data-collector to kiln_lib
