+++ kiln_lib/Cargo.toml
@@ -1,12 +1,12 @@
 [package]
-name = "data-collector"
+name = "kiln_lib"
 version = "0.1.0"
 edition = "2018"
 
 [dependencies]
-lambda_runtime = "0.2"
+chrono = "0.4"
 lambda_http = "0.1"
 http = "0.1"
-kiln_lib = { path = "../kiln_lib" }
 serde_json = "1.0"
 serde = { version = "1.0", features = ["derive"] }
+regex = "1.0"
